Created Crater and Depot autonomous.
Changed the isMoving detection in the drive chassis to try to speed up the
time between movement segments.

Crater will attempt to drive around the three minerals after sampling.
It will run collector in reverse once it gets to the end of the drive
plan to deposit the marker.  (assuming marker is in the collector).